+++ 甲状腺随访系统/model/HematologicalExamination .cs
@@ -11,77 +11,77 @@ namespace 甲状腺随访系统.MODEL
         /// <summary>
         /// （甲功）TSH
         /// </summary>
-        public float TSH { get; set; }
+        public Double TSH { get; set; }
 
         /// <summary>
         /// Tg洗脱液
         /// </summary>
-        public float lymphnodestg { get; set; }
+        public Double lymphnodestg { get; set; }
 
         /// <summary>
         /// ATG
         /// </summary>
-        public float ATG { get; set; }
+        public Double ATG { get; set; }
 
         /// <summary>
         /// CT
         /// </summary>
-        public float CT { get; set; }
+        public Double CT { get; set; }
 
         /// <summary>
         /// FT3
         /// </summary>
-        public float FT3 { get; set; }
+        public Double FT3 { get; set; }
 
         /// <summary>
         /// A-TPO
         /// </summary>
-        public float ATPO { get; set; }
+        public Double ATPO { get; set; }
 
         /// <summary>
         /// TG
         /// </summary>
-        public float TG { get; set; }
+        public Double TG { get; set; }
 
         /// <summary>
         /// FT4
         /// </summary>
-        public float FT4 { get; set; }
+        public Double FT4 { get; set; }
 
         /// <summary>
         /// PTH
         /// </summary>
-        public float PTH { get; set; }
+        public Double PTH { get; set; }
 
         /// <summary>
         /// CEA
         /// </summary>
-        public float CEA { get; set; }
+        public Double CEA { get; set; }
 
         /// <summary>
         ///（离子）K
         /// </summary>
-        public float K { get; set; }
+        public Double K { get; set; }
 
         /// <summary>
         /// Ca
         /// </summary>
-        public float Ca { get; set; }
+        public Double Ca { get; set; }
 
         /// <summary>
         /// Na
         /// </summary>
-        public float Na { get; set; }
+        public Double Na { get; set; }
 
         /// <summary>
         /// P
         /// </summary>
-        public float P { get; set; }
+        public Double P { get; set; }
 
         /// <summary>
         /// Cl
         /// </summary>
-        public float Cl { get; set; }
+        public Double Cl { get; set; }
 
     }
 }
