+++ 甲状腺随访系统/model/InspecAfterSurg.cs
@@ -10,42 +10,42 @@ namespace 甲状腺随访系统.MODEL
         /// <summary>
         /// 天数
         /// </summary>
-        public float days { get; set; }
+        public Double days { get; set; }
 
         /// <summary>
         /// PTH
         /// </summary>
-        public float PTH { get; set; }
+        public Double PTH { get; set; }
 
         /// <summary>
         /// Ca
         /// </summary>
-        public float Ca { get; set; }
+        public Double Ca { get; set; }
 
         /// <summary>
         /// P
         /// </summary>
-        public float P { get; set; }
+        public Double P { get; set; }
 
         /// <summary>
         /// 碱性磷酸酶
         /// </summary>
-        public float alkalinephosphatase { get; set; }
+        public Double alkalinephosphatase { get; set; }
 
         /// <summary>
         /// 25-强及维生素D
         /// </summary>
-        public float D { get; set; }
+        public Double D { get; set; }
 
         /// <summary>
         /// D2
         /// </summary>
-        public float D2 { get; set; }
+        public Double D2 { get; set; }
 
         /// <summary>
         /// D3
         /// </summary>
-        public float D3 { get; set; }
+        public Double D3 { get; set; }
