+++ 甲状腺随访系统/model/RadioIodine.cs
@@ -10,32 +10,32 @@ namespace 甲状腺随访系统.MODEL
         /// <summary>
         /// TS
         /// </summary>
-        public float TS { get; set; }
+        public Double TS { get; set; }
 
         /// <summary>
         /// FT3
         /// </summary>
-        public float FT3 { get; set; }
+        public Double FT3 { get; set; }
 
         /// <summary>
         /// FT4
         /// </summary>
-        public float FT4 { get; set; }
+        public Double FT4 { get; set; }
 
         /// <summary>
         ///sTG
         /// </summary>
-        public float sTG { get; set; }
+        public Double sTG { get; set; }
 
         /// <summary>
         /// ATG
         /// </summary>
-        public float ATG { get; set; }
+        public Double ATG { get; set; }
 
         /// <summary>
         /// 吸碘率（%）
         /// </summary>
-        public float iodineuptakerate { get; set; }
+        public Double iodineuptakerate { get; set; }
 
     }
 }
